QG3.jl

This package provides a Julia implementation of the Marshall, Molteni Quasigeostrophic Atmsopheric Model in three layers. It runs on CPU or CUDA GPUs and is differentiable via Zygote.jl.

The model is solved with a pseudo-spectral approach on a Gaussian grid with approatiate spherical harmonics transforms defined. Example scripts are provided in the examples folder.

Install e.g. via ]add https://github.com/maximilian-gelbrecht/QG3.jl.git and test the installation with ]test QG3

The repository includes pre-computed forcing and initial conditions on a T21 to run the model but no proper dataset to compute re-compute those in order to save space. The example folder also includes the necessary scripts to pre-copmute those for other grids and datasets.

The Model

Details about the model can be read up in "Towards a Dynamical Understanding of Planetary-Scale Flow Regimes", Marshall, Molteni, Journal of Atmsopheric Sciences, 1992.

Its governing equation for the quasigeostrophic vorticity $`q_i`$ in three equipressure levels (200, 500, 800 hPa) are given by

\[\dot{q_i} = -J(\psi_i, q_i) - D(\psi_i, q_i) + S \\ \vec{q} = T_{\psi q} \vec{\psi},\]

where the voriticy $q$ and streamfunction $\psi$ are related by a linear operator (comprising the Laplacian and temperature relaxation), $J$ is the Jacobian / advection term, $D$ the dissipation and $S$ a forcing computed from data.

Solvers

This packages includes all function necessary to define the governing equation of the QG3 Model. It does however, not include the differential equations solvers, for those we rely on DifferentialEquations.jl, i.e. OrdinaryDiffEq.jl. In the following, we will therefore briefly talk about how to define the right hand side necessary to define for those solvers, before going into a little more detail on the transforms and derivatives used throughout the model.

Initialize a model

The model parameter are hold as QG3ModelParameters:

QG3.QG3ModelParameters — Type

QG3ModelParameters{T}

Saves all the parameters of the model, also the units/normalization.

Fields

L::Int: number of SH number l, note that 0,...,lmax, so that L is lmax + 1

M::Int: # maximum number M values, M==2L - 1 / 2l_max + 1

N_lats::Int

N_lons::Int

lats::AbstractArray{T,1}

θ::AbstractArray{T,1} colatitudes = θ

μ::AbstractArray{T,1} sin(lats) == cos(colats)

LS::AbstractArray{T,2} Land see mask, on the same grid as lats and lons

h::AbstractArray{T,2}` orography, array on the same grid as lats and lons

R1i::T square inverse of Rossby radius 1 200-500hPa, default 1/(700km)^2

R2i::T square inverse of Rossby radius 2 500-800hPa, default 1/(450km)^2

H0::T scaling parameter for topography, default 9000m

τRi::T inverse of radiative time scale, default 1/(25d)

τEi::T inverse of drag time scale, default 1/(3d)

cH::T horizonatal diffusion coefficient

α1::T drag coefficient 1, default 0.5

α2::T drag coefficient 2, default 0.5

gridtype::String either 'gaussian' or 'regular', 'gaussian' leads to all transform being done by somewhat naively implemented transforms, 'regular' uses FastTransforms.jl

time_unit::T unit so that t [normalized] = t [d] / time_unit, default = 1 / 4π

distance_unit::T, default Earth's radius, so that s [normalized] = s [m] / distance_unit, default = 6.371e6

ψ_unit::T, derived from timeunit and distanceunit, so that ψ [normalized] = ψ [m^2/s] / ψ_unit

q_unit::T, derived from timeunit and distanceunit

Other constructors

QG3ModelParameters(L::Int, lats::AbstractArray{T,1}, lons::AbstractArray{T,1}, LS::AbstractArray{T,2}, h::AbstractArray{T,2}, R1i::T, R2i::T, H0::T, τRi::T, τEi::T, τHi::T, α1::T, α2::T, a::T, gridtype::String, time_unit::T, distance_unit::T)

Default arguments for most parameters, so that

QG3ModelParameters(L::Int, lats::AbstractArray{T,1}, lons::AbstractArray{T,1}, LS::AbstractArray{T,2}, h::AbstractArray{T,2})

works as well.

Normalization/units of the model

The model is normalized in (natural) units of the system.

Distance: The Earth's radius is set to 1 (and thus all units containing distance are scaled with the actual radius of the Earth)

Vorticity: Earth's angular speed is approximately 2π/d, here Earth's angular speed is set 2Ω = 1, so that the planetery vorticity component of the Jacobian is just ∂ψ/∂λ. Therefore: Ω = 1/2, it follows 2*2π [1/d] = 1 [Ωnormalized], therefore Ωnormalized = 4π/d.

Time: From the vorticity unit, follows the time unit is [d/4π] = [dnormalized]. Hence, time in days must be devided by 4π to get the natural time unit, or [dnormalized] = 4π [d]

both ψ and q thus need to expressed in these units too, this can be done using the ψ_unit and q_unit

For convenience, a pre-computed setup with forcing S, parameters qg3_pars and initial conditions ψ_0, q_0 can be loaded via

S, qg3_pars, ψ_0, q_0 = QG3.load_precomputed_data(GPU=false)

Alternatively, the pre-compute.jl script in the examples folder demonstrates how to compute those from other datasets and for other resolutions.

Then, the model can be pre-computed with the help of QG3Model

qg3 = QG3Model(qg3_pars)

or on GPU

qg3 = CUDA.@allowscalar QG3Model(qg3_pars) # on GPU the pre-computation need scalar indexing

These struct also hold the grid information qg3.g that comprises all transforms and derivatives.

GPU and CPU

Currently there are two different implementations, one that is optimised for CPU (the "2D version") and one that is optimised for GPU (the "3D version"). The GPU version can also run on CPU but not the other way around.

The GPU is automatically detected. If you wish to run the model on CPU on a computer that does have a CUDA capable GPU, then you can turn it of with QG3.gpuoff() before model pre-computation.

GPU Version

The GPU version is fully vectorized, without scalar indexing. All needed functions work on the 3d (level, lat, lon) / (level, il, m) field. The full problem definition thus is just simply

function QG3MM_gpu(q, m, t)
    p, S = m # parameters, forcing vector

    ψ = qprimetoψ(p, q)
    return - J(ψ, q, p) - D(ψ, q, p) + S
end

This right hand side calls the transform from potential vorticity to the streamfunction qprimetoψ, the advection operator J, the dissipation operator D and the pre-computed forcing S.

CPU Version

The CPU version works on 2d fields (lat, lon) / (il, m), on CPUs this is currently a little faster than using the GPU/3D version on CPU. In other words: the GPU version could probably still be optimized further. The problem definition follows as

function QG3MM_base(q, m, t)
    p, S = m # parameters, forcing vector

    ψ = qprimetoψ(p, q)
    return cat(
    reshape(- J(ψ[1,:,:], q[1,:,:], p) .- D1(ψ, q, p) .+ S[1,:,:], (1, p.p.L, p.p.M)),
    reshape(- J(ψ[2,:,:], q[2,:,:], p) .- D2(ψ, q, p) .+ S[2,:,:], (1, p.p.L, p.p.M)),
    reshape(- J3(ψ[3,:,:], q[3,:,:], p) .- D3(ψ, q, p) .+ S[3,:,:], (1, p.p.L, p.p.M)),
    dims=1)
end

Model Units

This implemenation uses a unit system, so that the Earth's radius and angular velocity equals

\[\begin{aligned} R = 1 \\ 2\Omega = 1, \end{aligned}\]

from these follow the units for all variables. The parameter struct holds those as qg3_pars.time_unit, qg3_pars.distance_unit, qg3_pars.ψ_unit and qg3_pars.q_unit.

Transforms

Throughout the model real spherical harmonics are used. The transform is implemented with a FFT and a Gauss-Legendre transform that is solved by pre-computing the associated Legendre polynomials and then executing the appropiate matrix multiplication for the transform. This ensures both differentiability and GPU compatability of the transforms. Gradients of the spherical harmonics transforms are checked against finite difference differentiation in the tests.

Transforms are called with transform_SH and transform_grid, e.g. as in the following:

ψ_0_grid = transform_grid(ψ_0, qg3)
ψ_0_SH = transform_SH(ψ_0_grid, qg3)

SHs coefficients on CPU are handled in the matrix convention that FastTransforms.jl uses: columns by m-value: 0, -1, 1, -2, 2, ..., rows l in ascending order. This was implemented initially during development to ensure compatability with FastTransforms.jl, unfortunately however, using FastTransforms.jl isn't currently working (yet) due to aliasing problems. On GPU, we store the coefficients differently to avoid additional operations to reorder all coefficents that would be needed otherwise with the CUFFT routines. Therefore the coefficient matrix convention is different on GPU, where the columns are ordered 0, 1, 2, .... l_max, 0, -1, -2, -3, .. .

Derivates

Whereas zonal derivatives and the Laplacian are computed in the SH domain, meridional derivatives are computed pseudo-spectrally with pre-computed values of the derivatives of the associated Legendre polynomials.

All derivatives follow a common naming scheme:

SHtoSH_dλ(ψ_0, qg3) # zonal derivative SH -> SH 
SHtoGrid_dλ(ψ_0, qg3) # zonal derivative SH -> Grid
SHtoGrid_dμ(ψ_0, qg3) # meridional (μ = sin(latitude)) derivative SH -> Grid 
SHtoGrid_dθ(ψ_0, qg3) # meridional (colatitude) derivative SH -> Grid 
Δ(ψ_0, qg3) # Laplacian SH -> SH

Transforms and Derivates for Machine Learning

Outside of the QG3 Model itself the transforms and derivative can also be used for ML tasks. For this purpose the constructors also allow to define them for more horizonatal levels and also for an additional (batch) dimension. The batch dimensions is added as a forth dimension, thus defining transforms and derivaties for fields (lvl x N_lat x N_lon x N_batch).

Either, the transforms and derivatives are initialized directly with the approatite N_level and N_batch input arguments via

QG3.GaussianGridtoSHTransform — Type

GaussianGridtoSHTransform(p::QG3ModelParameters{T}, N_level::Int=3; N_batch::Int=0)

Returns transform struct, that can be used with transform_SH. Transforms Gaussian Grid data to real spherical harmonics coefficients that follow the coefficient logic explained in the main documenation.

Additional input arguments:

N_level: defines the transform for N_level horizontal levels. Has to be equal to three for the QG3 model itself, but might be different for other applications.

N_batch: defines the transforms with an additional batch dimension for ML tasks, if N_batch==0 this is omitted

QG3.SHtoGaussianGridTransform — Type

SHtoGaussianGridTransform(p::QG3ModelParameters{T}, N_level::Int=3; N_batch::Int=0)

Returns transform struct, that can be used with transform_grid. Transforms real spherical harmonics coefficients to Gaussian grid data, follows the coefficient logic explained in the main documenation.

Additional input arguments:

N_level: defines the transform for N_level horizontal levels. Has to be equal to three for the QG3 model itself, but might be different for other applications.

N_batch: defines the transforms with an additional batch dimension for ML tasks, if N_batch==0 this is omitted

QG3.Derivative_dλ — Type

Derivative_dλ(p::QG3ModelParameters{T}; N_batch::Int=0)

Pre-computes Derivatives by longitude. Uses the SH relation, is therefore independ from the grid.

QG3.GaussianGrid_dμ — Type

GaussianGrid_dμ(p::QG3ModelParameters{T}, N_level::Int=3; N_batch::Int=0)

Pre-computes Pseudo-spectral approach to computing derivatives with repsect to μ = sin(lats). Derivatives are called with following the naming scheme: "Domain1Input"to"Domain2Output"_d"derivativeby"

QG3.Laplacian — Type

Laplacian(p::QG3ModelParameters{T}; init_inverse=false, R::T=T(1), N_batch::Int=0, kwargs...) where T

Initializes the Laplacian in spherical harmonics and if init_inverse==true also its inverse

Apply the Laplacian with the functions (@ref)[Δ] and (@ref)[Δ⁻¹]

or a they can be constructed all together via

QG3.grid — Function

grid(p::QG3ModelParameters{T}, gridtype::String, N_level::Int=3; N_batch::Int=0, kwargs...)

Convience constructor for the AbstractGridType based on the parameters set in p.