Reference

QG3.AbstractGridType — Type

AbstractGridType{T, onGPU}

Abstract type for grids. The grids save information about the transform from the spatial to spectral grid, e.g. pre-computed Legendre Polynomials

Required fields:

GtoSH

SHtoG

dμ

dλ

QG3.AbstractSHTransform — Type

abstract type AbstractSHTransform{onGPU}

Required fields for all subtypes:

outputsize

QG3.AbstractμDerivative — Type

abstract type AbstractμDerivative{onGPU} < AbstractDerivative{onGPU}

Required fields:

msinθ: To change between μ and latitude derivative (-sin(colats))

msinθ_3d: To change between μ and latitude derivative (-sin(colats))

QG3.ContinousTimeForcing — Type

TimeForcing

Time-dependent Forcing

QG3.Derivative_dλ — Type

Derivative_dλ(p::QG3ModelParameters{T}; N_batch::Int=0)

Pre-computes Derivatives by longitude. Uses the SH relation, is therefore independ from the grid.

QG3.GaussianGrid — Type

 GaussianGrid{T, onGPU} <: AbstractGridType{T, onGPU}

Struct for the transforms and derivates of a Gaussian Grid

Fields:

GtoSH::GaussianGridtoSHTransform

SHtoG::GaussianSHtoGridTransform

dμ::GaussianGrid_dμ

dλ::Derivative_dλ

Δ::Laplacian

∇8::Hyperdiffusion

QG3.GaussianGrid_dμ — Type

GaussianGrid_dμ(p::QG3ModelParameters{T}, N_level::Int=3; N_batch::Int=0)

Pre-computes Pseudo-spectral approach to computing derivatives with repsect to μ = sin(lats). Derivatives are called with following the naming scheme: "Domain1Input"to"Domain2Output"_d"derivativeby"

QG3.GaussianGridtoSHTransform — Type

GaussianGridtoSHTransform(p::QG3ModelParameters{T}, N_level::Int=3; N_batch::Int=0)

Returns transform struct, that can be used with transform_SH. Transforms Gaussian Grid data to real spherical harmonics coefficients that follow the coefficient logic explained in the main documenation.

Additional input arguments:

N_level: defines the transform for N_level horizontal levels. Has to be equal to three for the QG3 model itself, but might be different for other applications.

N_batch: defines the transforms with an additional batch dimension for ML tasks, if N_batch==0 this is omitted

QG3.Hyperdiffusion — Type

Hyperdiffusion(p::QG3ModelParameters{T}; scale::T=T(1)) where T

Initializes the Hyperdiffusion / horizonatal diffusion operator.

Apply it via the (@ref)[∇8] functions.

QG3.KineticEnergyCallback — Type

KineticEnergyCallback{T}(m::QG3Model)

Initializes a callback to compute the kinetic energy. Can be called with (u, t, integrator) and therefore used with a SavingCallback.

To initialize a SavingCallback you can e.g. do

vals = SavedValues(eltype(m), Vector{eltype(m)})
sol = solve(prob; cb=SavingCallback(KineticEnergyCallback(m), vals))
vals.saveval 

QG3.Laplacian — Type

Laplacian(p::QG3ModelParameters{T}; init_inverse=false, R::T=T(1), N_batch::Int=0, kwargs...) where T

Initializes the Laplacian in spherical harmonics and if init_inverse==true also its inverse

Apply the Laplacian with the functions (@ref)[Δ] and (@ref)[Δ⁻¹]

QG3.QG3Model — Type

QG3Model{T}

Holds all parameter and grid information, plus additional pre-computed fields that save computation time during model integration. All these parameter are seen as constants and not trainable. It should be computed on CPU, if you are on GPU use CUDA.@allowscalar QG3Model(...).

Fields

p::QG3ModelParameters{T} Parameters of the model

g::AbstractGridType{T} Grid type with pre-computed plans or Legendre polynomals and co.

k::AbstractArray{T,2} Array, drag coefficient pre-computed from orography and land-sea mask

TRcoeffs::AbstractArray{T,3} Array of Temperature relaxation coefficients (2d version)

TR_matrix Matrix for temperature relaxation (3d version)

cosϕ::AbstractArray{T,2} cos(latitude) pre computed

acosϕi::AbstractArray{T,2} inverse of a*cos(latitude) pre computed

Tψq matrix used for transforming stream function to voriticy q = Tψq * ψ + F

Tqψ`inverse of Tψq ψ = Tqψ*(q - F)

f modified coriolis vector used in transforming stream function to vorticity

f_J3 coriolis contribution to Jacobian at 850hPa

∂k∂ϕ derivates of drag coefficients, pre computed for Ekman dissipation

∂k∂μ

∂k∂λ includes 1/cos^2ϕ (for Ekman dissipiation computation)

QG3.QG3Model — Method

QG3Model(p::QG3ModelParameters; N_level::Integer=3, kwargs...)

Routine that pre computes the QG3 Model and returns a QG3Model struct with all precomputed fields except for the forcing.

The pre-computation is always done on CPU due to scalar indexing being used, if a GPU is avaible the final result is then transferred to the GPU.

If N_level` is set to values other than three, the pre-computations for the transforms and derivatives are computed with N_level levels, but the full model will not be working! This is just for accessing the transforms and derivatives

All $kwargs$ are forwarded to grid.

QG3.QG3ModelParameters — Type

QG3ModelParameters{T}

Saves all the parameters of the model, also the units/normalization.

Fields

L::Int: number of SH number l, note that 0,...,lmax, so that L is lmax + 1

M::Int: # maximum number M values, M==2L - 1 / 2l_max + 1

N_lats::Int

N_lons::Int

lats::AbstractArray{T,1}

θ::AbstractArray{T,1} colatitudes = θ

μ::AbstractArray{T,1} sin(lats) == cos(colats)

LS::AbstractArray{T,2} Land see mask, on the same grid as lats and lons

h::AbstractArray{T,2}` orography, array on the same grid as lats and lons

R1i::T square inverse of Rossby radius 1 200-500hPa, default 1/(700km)^2

R2i::T square inverse of Rossby radius 2 500-800hPa, default 1/(450km)^2

H0::T scaling parameter for topography, default 9000m

τRi::T inverse of radiative time scale, default 1/(25d)

τEi::T inverse of drag time scale, default 1/(3d)

cH::T horizonatal diffusion coefficient

α1::T drag coefficient 1, default 0.5

α2::T drag coefficient 2, default 0.5

gridtype::String either 'gaussian' or 'regular', 'gaussian' leads to all transform being done by somewhat naively implemented transforms, 'regular' uses FastTransforms.jl

time_unit::T unit so that t [normalized] = t [d] / time_unit, default = 1 / 4π

distance_unit::T, default Earth's radius, so that s [normalized] = s [m] / distance_unit, default = 6.371e6

ψ_unit::T, derived from timeunit and distanceunit, so that ψ [normalized] = ψ [m^2/s] / ψ_unit

q_unit::T, derived from timeunit and distanceunit

Other constructors

QG3ModelParameters(L::Int, lats::AbstractArray{T,1}, lons::AbstractArray{T,1}, LS::AbstractArray{T,2}, h::AbstractArray{T,2}, R1i::T, R2i::T, H0::T, τRi::T, τEi::T, τHi::T, α1::T, α2::T, a::T, gridtype::String, time_unit::T, distance_unit::T)

Default arguments for most parameters, so that

QG3ModelParameters(L::Int, lats::AbstractArray{T,1}, lons::AbstractArray{T,1}, LS::AbstractArray{T,2}, h::AbstractArray{T,2})

works as well.

Normalization/units of the model

The model is normalized in (natural) units of the system.

Distance: The Earth's radius is set to 1 (and thus all units containing distance are scaled with the actual radius of the Earth)

Vorticity: Earth's angular speed is approximately 2π/d, here Earth's angular speed is set 2Ω = 1, so that the planetery vorticity component of the Jacobian is just ∂ψ/∂λ. Therefore: Ω = 1/2, it follows 2*2π [1/d] = 1 [Ωnormalized], therefore Ωnormalized = 4π/d.

Time: From the vorticity unit, follows the time unit is [d/4π] = [dnormalized]. Hence, time in days must be devided by 4π to get the natural time unit, or [dnormalized] = 4π [d]

both ψ and q thus need to expressed in these units too, this can be done using the ψ_unit and q_unit

QG3.SHtoGaussianGridTransform — Type

SHtoGaussianGridTransform(p::QG3ModelParameters{T}, N_level::Int=3; N_batch::Int=0)

Returns transform struct, that can be used with transform_grid. Transforms real spherical harmonics coefficients to Gaussian grid data, follows the coefficient logic explained in the main documenation.

Additional input arguments:

N_level: defines the transform for N_level horizontal levels. Has to be equal to three for the QG3 model itself, but might be different for other applications.

N_batch: defines the transforms with an additional batch dimension for ML tasks, if N_batch==0 this is omitted

QG3.D — Method

D(ψ::AbstractArray{T,3}, q::AbstractArray{T,3}, m::QG3Model{T})

Computes the dissipiation of all levels, 850hPa has additional Ekman dissipation.

QG3.EK — Method

EK(ψ::AbstractArray{T,2}, m::QG3Model{T})

Computes the Ekman dissipation with:

EK = ∇(k∇ψ) = (∇k ∇ψ) + k Δψ EK = 1/a^2cos^2ϕ ∂k/∂λ ∂ψ/∂λ + 1/a^2 ∂k/∂ϕ ∂ψ/∂ϕ + k Δψ (a==1)

m.∂k∂λ includes 1/cos^2ϕ

QG3.H — Method

H(qprime::AbstractArray{T,3}, i::Int, m::QG3Model{T})

Computes the Horizontal diffusion at level i, q' is anomolous pv (without coriolis) 2D Fields m.cH∇8 = m.p.cH * m.∇8

QG3.H — Method

H(qprime::AbstractArray{T,3}, m::QG3Model{T})

Computes the horizontal diffusion, q' is anomolous pv (without coriolis) 3D Fields

QG3.J — Method

J(ψ::AbstractArray{T,3}, q::AbstractArray{T,3}, m::QG3Model{T})

Computes the Jacobian for the QG3 model in 3D Field formulation (e.g. for GPU). Takes as an imput the streamfunction ψ and the potential vorticity q' without the coriolis contribution. Adds the contribution of the Coriolis force and orography at the 850hPa level that is prescribed in the MM QG3 Model.

QG3.J — Method

J(ψ::AbstractArray{T,2}, q::AbstractArray{T,2}, g::AbstractGridType{T})

Compute the Jacobian determinant from ψ and q in μ,λ coordinates, J = ∂ψ/∂x ∂q/∂y - ∂ψ/∂y ∂q/∂x = 1/a^2cosϕ ( - ∂ψ/∂λ ∂q/∂ϕ + ∂ψ/∂ϕ ∂q/∂λ) = 1/a^2 (- ∂ψ/∂λ ∂q/∂μ + ∂ψ/∂μ ∂q/∂λ)

The last term ∂ψ/∂λ accounts for the planetery vorticity, actually it is 2Ω ∂ψ/∂λ, but 2Ω == 1, (write q = q' + 2Ωμ to proof it)

QG3.J3 — Method

J3(ψ::AbstractArray{T,2}, q::AbstractArray{T,2}, m::QG3Model{T})

For the Jacobian at 850hPa, q = q' + f(1+h/H0) = q' + f + f*h/H0, so that the thrid term has to be added.

QG3.J_F — Method

J_F(ψ::AbstractArray{T,N}, q::AbstractArray{T,N}, g::AbstractGridType{T})

Compute the Jacobian determinant from ψ and q in μ,λ coordinates without the planetary vorticity, as used in computing the eddy/transient forcing

QG3.J_F_Grid — Method

J_F_Grid(ψ::AbstractArray{T,N}, q::AbstractArray{T,N}, g::AbstractGridType{T})

Computes the Jacobian (without the planetary voriticity), input SPH, output Grid space

QG3.Jprime — Method

Jprime(ψ::AbstractArray{T,3}, q::AbstractArray{T,3}, g::AbstractGridType{T})

Computes the Jacobian (∂ψ/∂μ ∂q/∂λ - ∂ψ/∂λ ∂q/∂μ). Directly computes the derivatives of the two inputs ψ and q and doesn't account for any additional contribution (e.g. from the Coriolis force).

QG3.QG3MM_adv — Method

QG3MM_adv(q, p, t)

Just the Advection

dq = - J(ψ,q)

QG3.QG3MM_base — Method

QG3MM_base(q, p, t)

Base model used in the MM paper, with symmetrization around the equator

dq = - J(ψ,q) - D(ψ,q) + S

QG3.QG3MM_gpu — Method

QG3MM_gpu(q, p, t)

Base model used in the MM paper dq = - J(ψ,q) - D(ψ,q) + S

QG3.SH_zero_mask — Function

SH_zero_mask(p::QG3.QG3ModelParameters, size_tup=nothing)

Returns a mask that zeroes out all spurious elements of the SH coefficients + the (0,0) element which is also set zero

QG3.SHtoGrid_dθ — Method

derivative of input after θ (azimutal angle/colatitude) in SH, uses pre computed SH evaluations (dependend on the grid type)

QG3.SHtoGrid_dλ — Method

SHtoGrid_dλ(ψ, m::QG3Model{T})
SHtoGrid_dλ(ψ, m::AbstractGridType{T})

Derivative of input after φ (polar angle) or λ (longtitude) in SH to Grid

QG3.SHtoGrid_dμ — Method

SHtoGrid_dμ(ψ, m::QG3Model{T}) 
SHtoGrid_dμ(ψ, g::AbstractGridType) 
SHtoGrid_dμ(ψ, d::GaussianGrid_dμ)

Derivative of input after μ = sinϕ in SH, uses pre computed SH evaluations

QG3.SHtoGrid_dφ — Method

SHtoGrid_dφ(ψ, m::QG3Model{T})
SHtoGrid_dφ(ψ, m::AbstractGridType{T})
SHtoGrid_dφ(ψ, dλ::Derivative_dλ, sh2g::AbstractSHtoGridTransform)

Derivative of input after φ (polar angle) or λ (longtitude) in SH to Grid

QG3.SHtoGrid_dϕ — Method

SHtoGrid_dϕ(ψ, m)

Derivative of input after ϕ - latitude in SH, uses pre computed SH evaluations

QG3.SHtoSH_dλ — Method

SHtoSH_dλ(ψ, m::QG3Model{T})
SHtoSH_dλ(ψ, g::GaussianGrid)

Derivative of input after φ (polar angle/longtitude) in SH, output in SH

QG3.SHtoSH_dφ — Method

SHtoSH_dφ(ψ, m::QG3Model{T})
SHtoSH_dφ(ψ, g::GaussianGrid) 
SHtoSH_dφ(ψ, g::Derivative_dλ)

Derivative of input after φ (polar angle/longtitude) in SH, output in SH

QG3.TR — Method

TR(ψ::AbstractArray{T,3}, m::QG3Model{T})

Computes the temperature relaxation of all levels

QG3.TR — Method

TR(m::QG3Model{T}, ψ1, ψ2, Ri::T)

Computes the Temperature relaxation between layers with streamfunction ψ1 and ψ2.

QG3.change_msign! — Method

change_msign!(A)

Change the sign of the m in SH (FastTranforms.jl convention of storing them). This version swaps the columns inplace

QG3.change_msign — Method

change_msign(A)

Change the sign of the m in SH. This version returns a view

there is currently a bug or at least missing feature in Zygote, the AD library, that stops views from always working flawlessly when a view is mixed with prior indexing of an array. We need a view for the derivative after φ to change the sign of m, so here is a differentiable variant of the SHtoSH_dφ function for the 2d field

QG3.compute_GaussWeights — Method

Pre-computes gaussian weights for Legendre Transform, also checks if we really have the correct gaussian latitudes

QG3.compute_P — Method

Pre-compute ass. Legendre Polynomials and dP/dx (derivative of ass. Legendre Polynomial) at the grid points and also the remainder of the Spherical Harmonics at the grid points using GSL

m values are stored 0,-1,1,-2,2,-3,3,... (on CPU) m values are stored 0,1,2,3,4,5,6,7, ...l_max, 0 (nothing),-1, -2, -3, (on GPU) (the second 0 is the Imanigary part / sin part of the fourier transform which is always identical to zero, it is kept here to have equal matrix sizes)

so far only |m| is used, as I assume real SPH.

QG3.compute_S_Roads — Method

compute_S_Roads(data::AbstractArray{T,4}, m::QG3Model{T}; datasource="ψ")

Pre-compute Forcing S data level,X,Y,t

Compute the forcing according to Roads, this is outlined in the MM and Corti paper as well.

This uses an unfiltered version, in the Corti paper there is a filter. Maybe implement it as well

data, input data

m, QG3Model Parameters and pre-computed fields

kwargs

datasource, either 'qprime' (anomalous potential vorticity) or 'sf' (streamfunction)

QG3.compute_S_t_corti — Function

Pre-compute time-dependent Forcing S data level,X,Y,t

Like in the Corti paper for the seasonal run this interpolates between two extremal forcings.

Nt_year: how many days does a year have / period length of the sinus

QG3.compute_TR — Method

Pre-compute array of temperature relaxation coefficients.

(l=0, m=0)-coefficient is assigned zero. This results in this expansion coefficient beeing constant in the whole model.

QG3.compute_TR_matrix — Method

Pre-compute matrices involved in the temperature relaxation in matrix form (for GPU / 3d fields)

TR = TRMatrix * ψ ∀ l ∈ [0,lmax]

For l=0 the matrix is set to be zero

QG3.compute_acosϕi — Method

Pre-compute cos(ϕ)^-1 (latitude matrix)

QG3.compute_batched_TR_matrix — Method

compute_batched_TR_matrix

prepares the TR = TRMatrix * ψ ∀ l ∈ [0,lmax] matrix for batched multiply

QG3.compute_batched_ψq_transform_matrices — Method

compute_batched_ψq_transform_matrices

prepares the transform from q to ψ and back for a batched matrix vector multiply, see also (@ref)[compute_ψq_transform_matrices]

QG3.compute_coriolis_vector_grid — Method

Pre-compute a matrix with with the Coriolis factor

QG3.compute_cosϕ — Method

Pre-compute cos(ϕ) (latitude) matrix

QG3.compute_f_J3 — Method

Pre-compute the additional contribution of the Coriolis force to the 850hPa Jacobian component in GPU ready format, for the Jacobian at 850hPa, q = q' + f(1+h/H0) = q' + f + f*h/H0, so that the thrid term has to be added.

QG3.compute_k — Method

Pre-compute drag coefficient k in real grid space, double check with values from paper

QG3.compute_Δ — Method

Pre-compute the Laplacian in Spherical Harmonics, follows the matrix convention of FastTransforms.jl

QG3.compute_Δ⁻¹ — Method

Pre-compute the inverse Laplacian in Spherical Harmonics, follows the matrix convention of FastTransforms.jl

QG3.compute_ψq_transform_matrices — Method

Pre-compute matrices involved in transforming vorticity to streamfunction and back

q = Tψq * ψ + f ψ = Tqψ * (q - f) = Tψq^(-1) * (q - f)

For l=0 the matrix is singular. In the fortran code they set q to something which for me does not make sense. Here Tψq(l=0) is just 1, so that transforming back and forth recovers the correct ψ. This makes sense in my eyes because q is only ever used for its spatial derivatives, all of which are 0 for l=0.

QG3.compute_∇8 — Method

" Pre-compute the 8-th derivative in Spherical Harmonics

QG3.gpuoff — Method

gpuoff()

Manually toggle GPU use off

QG3.gpuon — Method

gpuon()

Manually toggle GPU use on (if available)

QG3.grid — Method

grid(p::QG3ModelParameters{T}, gridtype::String, N_level::Int=3; N_batch::Int=0, kwargs...)

Convience constructor for the AbstractGridType based on the parameters set in p.

QG3.hours — Method

hours(x, m::QG3Model)

Convert from model time units to hours

QG3.isongpu — Method

isongpu(m::QG3Model{T})

Determines if the model was pre-computed to be used on GPU.

QG3.kinetic_energy — Method

kinetic_energy(ψ::AbstractArray{T,3}, m::QG3Model{T}) where T

Computes the kinetic energy per layer

QG3.lMatrix — Method

lMatrix(L, M; GPU=nothing)

Return l-Matrix of SH coefficients in convention of FastTransforms.jl.

QG3.level_index — Method

level_index(strings,lvls)

Helper function for indexing netcdf files with many levels.

string, array with name of levels to be selected

lvls, array with all levels

QG3.load_precomputed_data — Method

load_precomputed_data(GPU=false)

Loads the precomputed data that is saved in the package. It is computed from ERA5 T21 u/v data. Returns in order

S, qg3ppars, ψ_0, q_0

Forcing, Parameters, Streamfunction initial conditions, vorticity initial conditions

If GPU==true returns those in GPU SPH order as CuArrays, otherwise uses CPU SPH order.

QG3.mMatrix — Method

mMatrix(L::Integer, M::Integer)

Pre-compute a matrix with (m) values of the SH matrix format of FastTransforms.jl, used for zonal derivative

Kwarg 'GPU', if given, overrides the automatic detection of wheather or not a GPU is avaible.

QG3.make3d — Method

make3d(A::AbstractArray{T,2})

repeats an array three times to turn it into (3 x size(A,1) x size(A,2)), for the fully matrix version of model.

QG3.qprimetoψ — Method

qprimetoψ(p::QG3Model{T}, q::AbstractArray{T,3})

Convert the anomalous potential vorticity q' to streamfunction ψ.

This version is slightly slower than the old one (as it not aware of the matrix being half full of zeroes), but it is non-mutating which makes it suitable for the automatic differentation.

It replaces the double loop over the coefficient matrix with a batched vector multiply. The advantage of that is that it is optimised for GPU, so it might actually be faster on the GPU than doing a manual loop.

QG3.qprimetoψ — Method

qprimetoψ(p::QG3Model{T}, q::AbstractArray{T,4})

Convert the anomalous potential vorticity q' to streamfunction ψ.

QG3.qtoψ — Method

qtoψ(p::QG3Model{T}, q::AbstractArray{T,3})

Convert the potential vorticity q to streamfunction ψ. See also qprimetoψ.

QG3.reorder_SH_cpu — Method

reorder_SH_cpu(A::AbstractArray{T,3},p::QG3ModelParameters)

Reorders the SH coefficient so that computations on CPU are more efficient, inverse of reorder_SH_gpu

QG3.reorder_SH_cpu — Method

reorder_SH_cpu(A::AbstractArray{T,4},p::QG3ModelParameters)

Reorders the SH coefficient so that computations on CPU are more efficient, inverse of reorder_SH_gpu

QG3.reorder_SH_cpu — Method

reorder_SH_cpu(A::AbstractArray{T,2},p::QG3ModelParameters)

Reorders the SH coefficient so that computations on CPU are more efficient, inverse of reorder_SH_gpu

QG3.reorder_SH_gpu — Method

Reorders the SH coefficient so that computations on GPU are more efficient, it also includes the truncation as the SH array has size Nlat x Nlon, however all entries outside of L x M are zeroes.

Original order (FastTransforms.jl) columns by m : 0, -1, 1, -2, -2, ....

New order columns by m: 0, 1, 2, ... l_max, 0 (nothing), -1, -2, ..

2d input assumes L x M matrix, (all fields in SH), also enlarges the matrix to include truncation (additional elements are zero) to Nlat x Nlons 3d input assumes Nlat (or 3) x L x M matrix (e.g. precomputed legendre polynomials), also enlarges the matrix to include truncation (additional elements are zero) to Nlat (or 3) x Nlats x Nlons

Incase CUDA is not used, it just return the input.

QG3.sph_zero_spurious_modes! — Method

Sets the unphysical numbers zero (usually only needed for FastTransforms.jl)

QG3.transform_SH — Function

transform_SH(data::AbstractArray{T,N}, t::GaussianGridtoSHTransform) 
transform_SH(data::AbstractArray, m::QG3Model)

Transforms data into the spherical harmonics domain. The coefficents are ordered in a matrix in coloumns of the m value. On CPU the convention of FastTransform.jl is used (0, -1, 1, -2, 2, ...), on GPU the convention (0, 1, 2, 3, ...., (nothing, -1, -2, -3, ...)). Watch out, in future proabaly this might be standardized.

QG3.transform_grid — Function

transform_grid(data::AbstractArray{T,N}, t::SHtoGaussianGridTransform)

Transforms data from the spherical harmonics domain to a Gaussian Grid. The coefficents are ordered in a matrix in coloumns of the m value. On CPU the convention of FastTransform.jl is used (0, -1, 1, -2, 2, ...), on GPU the convention (0, 1, 2, 3, ...., (nothing, -1, -2, -3, ...)). Watch out, in future proabaly this might be standardized.

transform_grid(data, m::QG3Model; varname="ψ")

If varname ∈ ["pv","q","vorticity"] the [1,1] element is set to zero first, as this element is set to the streamfunction-[1,1] usually in the ψtoq routines.

QG3.zeros_Grid — Method

zeros_Grid(p::QG3Model{T})

Returns a zero array in the dimensions of the grid.

QG3.zeros_SH — Method

zeros_SH(p::QG3Model{T})

Returns a zero array in the dimensions of the SH.

QG3.Δ — Method

Δ(ψ::AbstractArray, L::Laplacian{T})
Δ(ψ::AbstractArray, m::QG3Model{T})

Apply the Laplacian. Also serves to convert regular vorticity (not the quasigeostrophic one) to streamfunction)

QG3.Δ⁻¹ — Method

Δ⁻¹(ψ::AbstractArray, m::QG3Model{T})

Apply the inverse Laplacian. Also serves to convert the streamfunction to regular vorticity

QG3.ψtoq — Method

ψtoq(p::QG3Model{T}, ψ::AbstractArray{T,3})

Convert the streamfunction ψ to the potential vorticity. See also ψtoqprime.

QG3.ψtoqprime — Method

ψtoqprime(p::QG3Model{T}, ψ::AbstractArray{T,3})

Convert the streamfunction ψ to (anomlous) potential vorticity q' in spherical harmonics basis. See also ψtoq.

This version is slightly slower than the old one on CPU (as it not aware of the matrix being half full of zeroes), but it is non-mutating which makes it suitable for the automatic differentation.

It replaces the double loop over the coefficient matrix with a batched vector multiply. The advantage other besides it being non-mutating is that it is optimised for GPU, so it might actually be faster on the GPU than doing a manual loop.

QG3.ψtoqprime — Method

ψtoqprime(p::QG3Model{T}, ψ::AbstractArray{T,4})

Convert the anomalous potential vorticity q' to streamfunction ψ.

QG3.∇8 — Method

∇8(q::AbstractArray{T,N}, H::Hyperdiffusion{T})

Apply the hyperdiffusion to the input