QG3.AbstractGridType — Type

AbstractGridType{T, onGPU}

Abstract type for grids. The grids save information about the transform from the spatial to spectral grid, e.g. pre-computed Legendre Polynomials

QG3.ContinousTimeForcing — Type

TimeForcing

Time-dependent Forcing

QG3.QG3Model — Type

QG3Model{T}

Holds all parameter and grid information, plus additional pre-computed fields that save computation time during model integration

QG3.QG3Model — Method

QG3Model(p::QG3ModelParameters)

Routine that pre computes the QG3 Model and returns a QG3Model struct with all precomputed fields except for the forcing.

QG3.QG3ModelParameters — Type

QG3ModelParameters{T}

Saves all the parameters of the model

Fields

* `L::Int`: number of SH number l, note that 0,...,l_max, so that L is l_max + 1
* `M::Int`: # maximum number M values, M==2*L - 1 / 2*l_max + 1
* `N_lats::Int`
* `N_lons::Int`
* `lats::AbstractArray{T,1}`
* `colats::AbstractArray{T,1}` colatitudes = θ
* `μ::AbstractArray{T,1}` sin(lats) == cos(colats)
* `LS::AbstractArray{T,2}` Land see mask, on the same grid as lats and lons
* `h::AbstractArray{T,2}`` orography, array on the same grid as lats and lons
* `R1i::T` square inverse of Rossby radius 1 200-500hPa, default 1/(700km)^2
* `R2i::T` square inverse of Rossby radius 2 500-800hPa, default 1/(450km)^2
* `H0::T` scaling parameter for topography, default 9000m
* `τRi::T` inverse of radiative time scale, default 1/(25d)
* `τEi::T` inverse of drag time scale, default 1/(3d)
* `cH::T` horizonatal diffusion coefficient
* `α1::T` drag coefficient 1, default 0.5
* `α2::T` drag coefficient 2, default 0.5
* `a::T` Earth's radius  # WIP: some functions assume a=1
* `Ω::T` Earth's angular speed for Coriolis # WIP: some functions assume Ω=1/2
* `gridtype::String` either 'gaussian' or 'regular', 'gaussian' leads to all transform being done by somewhat naively implemented transforms, 'regular' uses FastTransforms.jl
* `time_unit::T` unit so that t [normalized] = t [s] / time_unit, default = 1/4π
* `distance_unit::T`, default Earth's radius
* `ψ_unit::T`, derived from time_unit and distance_unit
* `q_unit::T`, derived from time_unit and distance_unit

Other constructors

QG3ModelParameters(L::Int, lats::AbstractArray{T,1}, lons::AbstractArray{T,1}, LS::AbstractArray{T,2}, h::AbstractArray{T,2}, R1i::T, R2i::T, H0::T, τRi::T, τEi::T, τHi::T, α1::T, α2::T, a::T, Ω::T, gridtype::String, time_unit::T, distance_unit::T)

Default arguments for most parameters, so that

QG3ModelParameters(L::Int, lats::AbstractArray{T,1}, lons::AbstractArray{T,1}, LS::AbstractArray{T,2}, h::AbstractArray{T,2})

works as well.

QG3.EK — Method

Ekman dissipation

EK = ∇(k∇ψ) = (∇k ∇ψ) + k Δψ EK = 1/a^2cos^2ϕ ∂k/∂λ ∂ψ/∂λ + 1/a^2 ∂k/∂ϕ ∂ψ/∂ϕ + k Δψ (a==1)

m.∂k∂λ includes 1/cos^2ϕ

QG3.EK3_simple — Method

Simplified Ekman dissipiation for k(ϕ,λ) = const

QG3.H — Method

Horizontal diffusion, q' is anomolous pv (without coriolis)

QG3.J — Method

Compute the Jacobian determinant from ψ and q in μ,λ coordinates, J = ∂ψ/∂x ∂q/∂y - ∂ψ/∂y ∂q/∂x = 1/a^2cosϕ ( - ∂ψ/∂λ ∂q/∂ϕ + ∂ψ/∂ϕ ∂q/∂λ) = 1/a^2 (- ∂ψ/∂λ ∂q/∂μ + ∂ψ/∂μ ∂q/∂λ)

The last term ∂ψ/∂λ accounts for the planetery vorticity, actually it is 2Ω ∂ψ/∂λ, but 2Ω == 1, (write q = q' + 2Ωμ to proof it)

QG3.J3 — Method

For the Jacobian at 850hPa, q = q' + f(1+h/H0) = q' + f + f*h/H0, so that the thrid term has to be added.

QG3.J_F — Method

Compute the Jacobian determinant from ψ and q in μ,λ coordinates without the planetary vorticity, as used in computing the eddy/transient forcing

QG3.QG3MM_adv — Method

QG3MM_adv(q, p, t)

Just the Advection

dq = - J(ψ,q)

QG3.QG3MM_base — Method

QG3MM_base(q, p, t)

Base model used in the MM paper, with symmetrization around the equator

dq = - J(ψ,q) - D(ψ,q) + S

QG3.SHtoGrid_dθ — Method

derivative of input after θ (azimutal angle/colatitude) in SH, uses pre computed SH evaluations (dependend on the grid type)

QG3.SHtoGrid_dμ — Method

derivative of input after μ = sinϕ in SH, uses pre computed SH evaluations

QG3.SHtoGrid_dφ — Method

derivative of input after φ (polar angle) or λ (longtitude) in SH to Grid, only for a single layer

QG3.SHtoSH_dφ — Method

derivative of input after φ (polar angle/longtitude) in SH, output in SH

these are variants with AbstractArray{T,3} and index i to select which layer is the input for the derivative.

there is currently a bug or at least missing feature in Zygote, the AD library, that stops views from always working flawlessly when a view is mixed with prior indexing of an array. We need a view for the derivative after φ to change the sign of m, so here is a differentiable variant of the SHtoSH_dφ function

QG3.SHtoSH_dφ — Method

derivative of input after φ (polar angle/longtitude) in SH, output in SH

QG3.TR — Method

Temperature relaxation

QG3._SHtoGrid_dμθ — Method

Performs the latitude-based derivates. Uses a form of synthesis based on pre-computed values of the ass. Legendre polynomials at the grid points.

This version is optimized to be non-mutating with batchedvec. The inner batchedvec correspodends to an inverse Legendre transform and the outer multiply to an inverse Fourier transform.

QG3.change_msign! — Method

change_msign(A)

Change the sign of the m in SH (FastTranforms.jl convention of storing them). This version swaps the columns inplace

QG3.change_msign — Method

change_msign(A)

Change the sign of the m in SH (FastTranforms.jl convention of storing them). This version returns a view

QG3.compute_GaussWeights — Method

Pre-computes gaussian weights for Legendre Transform, also checks if we really have the correct gaussian latitudes

QG3.compute_P — Method

Pre-compute ass. Legendre Polynomials and dP/dx (derivative of ass. Legendre Polynomial) at the grid points and also the remainder of the Spherical Harmonics at the grid points using GSL

m values are stored 0,-1,1,-2,2,-3,3,...

double check -m values , so far only |m| is used, as I assume real SPH.

QG3.compute_S_Roads — Method

Pre-compute Forcing S data level,X,Y,t

Compute the forcing according to Roads, this is outlined in the MM and Corti paper as well.

This uses an unfiltered version, in the Corti paper there is a filter. Maybe implement it as well

data, input data

m, QG3Model Parameters and pre-computed fields

kwargs

datasource, either 'qprime' (anomalous potential vorticity) or 'sf' (streamfunction)

QG3.compute_S_t_corti — Function

Pre-compute time-dependent Forcing S data level,X,Y,t

Like in the Corti paper for the seasonal run this interpolates between two extremal forcings.

Nt_year: how many days does a year have / period length of the sinus

QG3.compute_TR — Method

Pre-compute array of temperature relaxation coefficients.

(l=0, m=0)-coefficient is assigned zero. This results in this expansion coefficient beeing constant in the whole model.

QG3.compute_acosϕi — Method

Pre-compute (a*cos(ϕ))^-1 (latitude matrix)

QG3.compute_batched_ψq_transform_matrices — Method

compute_batched_ψq_transform_matrices

prepares the transform from q to ψ and back for a batched matrix vector multiply, see also (@ref)[compute_ψq_transform_matrices]

QG3.compute_coriolis_vector_SH — Method

Pre-compute a matrix with with the Coriolis factor

QG3.compute_cosϕ — Method

Pre-compute cos(ϕ) (latitude) matrix

QG3.compute_k — Method

Pre-compute drag coefficient k in real grid space, double check with values from paper

QG3.compute_mmMatrix — Method

Pre-compute a matrix with (m) values of the SH matrix format of FastTransforms.jl, used for zonal derivative

QG3.compute_Δ — Method

Pre-compute the Laplacian in Spherical Harmonics, follows the matrix convention of FastTransforms.jl

QG3.compute_ψq_transform_matrices — Method

Pre-compute matrices involved in transforming vorticity to streamfunction and back

q = Tψq * ψ + f ψ = Tqψ * (q - f) = Tψq^(-1) * (q - f)

For l=0 the matrix is singular. In the fortran code they set q to something which for me does not make sense. Here Tψq(l=0) is just 1, so that transforming back and forth recovers the correct ψ. This makes sense in my eyes because q is only ever used for its spatial derivatives, all of which are 0 for l=0.

QG3.compute_∇8 — Method

" Pre-compute the 8-th derivative in Spherical Harmonics

QG3.lMatrix — Method

Return l-Matrix of SH coefficients in convention of FastTransforms.jl

QG3.level_index — Method

level_index(strings,lvls)

Helper function for indexing netcdf files with many levels.

string, array with name of levels to be selected

lvls, array with all levels

QG3.qprimetoψ — Method

Convert the anomalous potential vorticity q' to streamfunction ψ

This version is slightly slower than the old one (as it not aware of the matrix being half full of zeroes), but it is non-mutating which makes it suitable for the automatic differentation.

It replaces the double loop over the coefficient matrix with a batched vector multiply. The advantage of that is that it is optimised for GPU, so it might actually be faster on the GPU than doing a manual loop.

QG3.qtoψ — Method

Convert the potential vorticity q to streamfunction ψ

QG3.sph_zero_spurious_modes! — Method

Sets the unphysical numbers zero (usually only needed for FastTransforms.jl)

QG3.transformGGridtoSH — Method

manually implemented transfrom from grid to spherical space, so far only for gaussian grid

CPU variant

QG3.transformGGridtoSH — Method

manually implemented transfrom from grid to spherical space, so far only for gaussian grid

GPU variant, r2c fft instead of r2r fft

QG3.transformSHtoGGrid — Method

manually implemented transform from spherical harmonics to grid. This works for all grid types, the grid type is indirectly specified through the pre-computed ass. legendre polynomials

CPU variant

QG3.transform_SH — Method

transform_SH(data, m::QG3Model, varname::String="ψ", setzeros=true)

Transforms the data to spherical harmonics.

if setzeros == true, the "unphysical" lower triangle is directly set zero when using FastTransforms.jl

QG3.transform_grid — Method

transform_grid(data, m::QG3Model)

Transforms the data to real space

If varname ∈ ["pv","q","vorticity"] the [1,1] element is set to zero first, as this element is set to the streamfunction-[1,1] usually in the ψtoq routines.

QG3.ψtoq — Method

Convert the streamfunction ψ to the potential vorticity

QG3.ψtoqprime — Method

Convert the streamfunction ψ to (anomlous) potential vorticity q' in spherical harmonics basis

This version is slightly slower than the old one on CPU (as it not aware of the matrix being half full of zeroes), but it is non-mutating which makes it suitable for the automatic differentation.

It replaces the double loop over the coefficient matrix with a batched vector multiply. The advantage other besides it being non-mutating is that it is optimised for GPU, so it might actually be faster on the GPU than doing a manual loop.